nexum-backend/worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

# ML
torch==2.3.1
torchvision==0.18.1
transformers==4.44.2
faiss-cpu==1.8.0
Pillow==10.4.0
imagehash==4.3.1
numpy==1.26.4

# GCP
google-cloud-storage==2.18.2
google-cloud-firestore==2.17.2
google-cloud-pubsub==2.23.0

# Utils
python-multipart==0.0.9
httpx==0.27.2
structlog==24.4.0
